+++ lib/ui/messages/MessagesDirectoryPanel.dart
@@ -2,14 +2,17 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/ui/messages/MessagesConversationPanel.dart';
 import 'package:illinois/ui/messages/MessagesWidgets.dart';
+import 'package:illinois/ui/profile/ProfileDirectoryAccountsList.dart';
 import 'package:illinois/ui/profile/ProfileDirectoryAccountsPage.dart';
 import 'package:illinois/ui/profile/ProfileDirectoryPage.dart';
 import 'package:illinois/ui/profile/ProfileDirectoryWidgets.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/TabBar.dart' as uiuc;
 import 'package:illinois/utils/AppUtils.dart';
+import 'package:rokwire_plugin/model/auth2.directory.dart';
 import 'package:rokwire_plugin/model/social.dart';
 import 'package:rokwire_plugin/service/localization.dart';
+import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/social.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
@@ -25,11 +28,12 @@ class MessagesDirectoryPanel extends StatefulWidget {
   _MessagesDirectoryPanelState createState() => _MessagesDirectoryPanelState();
 }
 
-class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with TickerProviderStateMixin {
+class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with TickerProviderStateMixin implements NotificationsListener {
   final GlobalKey<RecentConversationsPageState> _recentPageKey = GlobalKey();
-  final GlobalKey<ProfileDirectoryAccountsPageState> _allUsersPageKey = GlobalKey();
+  GlobalKey<ProfileDirectoryAccountsPageState> _allUsersPageKey = GlobalKey();
   final ScrollController _recentScrollController = ScrollController();
   final ScrollController _allUsersScrollController = ScrollController();
+
   late TabController _tabController;
   int _selectedTab = 0;
 
@@ -38,8 +42,18 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
     Localization().getStringEx('panel.messages.directory.tab.all.label', 'All Users'),
   ];
 
+  String _searchText = '';
+  Map<String, dynamic> _filterAttributes = <String, dynamic>{};
+
+  final Set<String> _selectedAccountIds = <String>{};
+
   @override
   void initState() {
+    NotificationService().subscribe(this, [
+      Social.notifyConversationsUpdated,
+      Social.notifyMessageSent,
+    ]);
+
     _tabController = TabController(length: _tabNames.length, initialIndex: _selectedTab, vsync: this);
     _tabController.addListener(_onTabChanged);
 
@@ -48,12 +62,30 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
 
   @override
   void dispose() {
+    NotificationService().unsubscribe(this);
+
     _tabController.removeListener(_onTabChanged);
     _tabController.dispose();
+
     _recentScrollController.dispose();
     _allUsersScrollController.dispose();
 
     super.dispose();
+  }
+
+  // NotificationsListener
+  @override
+  void onNotification(String name, dynamic param) {
+    if (name == Social.notifyConversationsUpdated) {
+      if (mounted) {
+        setState(() {
+          _selectedAccountIds.clear();
+        });
+      }
+    }
+    else if (name == Social.notifyMessageSent) {
+      setState(() {});
+    }
   }
 
   @override
@@ -74,6 +106,10 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
           Localization().getStringEx('panel.messages.directory.header.message', 'Who do you want to send to?'),
           style: Styles().textStyles.getTextStyle('widget.title.large.fat')
         ),
+      ),
+      Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 16.0),
+        child: _searchBarWidget,
       ),
       Padding(
         padding: const EdgeInsets.symmetric(horizontal: 16.0),
@@ -114,6 +150,7 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
     return tabs;
   }
 
+
   Widget _buildContinueButtonOverlay(Widget content, { ScrollController? scrollController }) {
     return RefreshIndicator(onRefresh: _onRefresh, child:
       Stack(
@@ -148,43 +185,65 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
   Widget get _recentContent =>
     RecentConversationsPage(
       key: _recentPageKey,
+      searchText: _searchText,
       recentConversations: widget.recentConversations,
       conversationPageSize: widget.conversationPageSize,
       scrollController: _recentScrollController,
-      onSelectedConversationChanged: _onSelectedConversationChanged,
+      onConversationSelectionChanged: _onConversationSelectionChanged,
+      selectedAccountIds: _selectedAccountIds,
     );
 
   Widget get _allUsersContent =>
-    ProfileDirectoryAccountsPage(DirectoryAccounts.userDirectory,
-      displayMode: DirectoryDisplayMode.select,
+    ProfileDirectoryAccountsList(DirectoryAccounts.userDirectory,
       key: _allUsersPageKey,
+      displayMode: DirectoryDisplayMode.select,
       scrollController: _allUsersScrollController,
-      onSelectedAccountsChanged: _onSelectedAccountsChanged,
+      searchText: _searchText,
+      filterAttributes: _filterAttributes,
+      onAccountSelectionChanged: _onAccountSelectionChanged,
+      selectedAccountIds: _selectedAccountIds,
     );
+
+  // Search & Filters
+
+  Widget get _searchBarWidget =>
+    ProfileDirectoryFilterBar(
+      key: ValueKey(ProfileDirectoryFilter(searchText: _searchText, attributes: _filterAttributes)),
+      searchText: _searchText,
+      onSearchText: _onSearchText,
+      filterAttributes: (_selectedTab != 0) ? _filterAttributes : null,
+      onFilterAttributes: _onFilterAttributes,
+    );
+
+  void _onSearchText(String text) {
+    setStateIfMounted((){
+      _searchText = text;
+      _allUsersPageKey = GlobalKey();
+    });
+  }
+
+  void _onFilterAttributes(Map<String, dynamic> filterAttributes) {
+    setStateIfMounted((){
+      _filterAttributes = filterAttributes;
+      _allUsersPageKey = GlobalKey();
+    });
+  }
+
+  // Event Handlers
 
   Future<void> _onTapCreateConversation() async {
     // do not need to check for existing conversation with selected members because Social BB handles it
-    Set<String>? memberIds = (_selectedTab == 0) ?
-      _recentPageKey.currentState?.selectedAccountIds :
-      _allUsersPageKey.currentState?.selectedAccountIds;
-
-    if ((memberIds != null) && memberIds.isNotEmpty) {
-      Conversation? conversation = await Social().createConversation(memberIds: memberIds.toList());
+    if (_hasSelectedAccounts) {
+      Conversation? conversation = await Social().createConversation(memberIds: _selectedAccountIds.toList());
       if (mounted) {
         if (conversation != null) {
-          if (_selectedTab == 0) {
-            _allUsersPageKey.currentState?.clearSelectedIds();
-          }
-          else {
-            _recentPageKey.currentState?.clearSelectedIds();
-          }
+          clearSelectedIds();
           Navigator.push(context, CupertinoPageRoute(builder: (context) => MessagesConversationPanel(conversation: conversation,)));
         } else {
           AppAlert.showDialogResult(context, Localization().getStringEx('panel.messages.directory.button.continue.failed.msg', 'Failed to create a conversation with the selected members.'));
         }
       }
     }
-
   }
 
   void _onTabChanged({bool manual = true}) {
@@ -195,20 +254,38 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
     }
   }
 
-  void _onSelectedAccountsChanged() {
-    setStateIfMounted(() {});
+  void _onAccountSelectionChanged(Auth2PublicAccount account, bool value) {
+    if (account.id?.isNotEmpty ?? false) {
+      setStateIfMounted(() {
+        if (value) {
+          _selectedAccountIds.add(account.id!);
+        } else {
+          _selectedAccountIds.remove(account.id);
+        }
+      });
+    }
   }
 
-  void _onSelectedConversationChanged() {
-    setStateIfMounted(() {});
+  void _onConversationSelectionChanged(bool value, Conversation conversation) {
+    setStateIfMounted(() {
+      if (value) {
+        _selectedAccountIds.addAll(conversation.memberIds ?? []);
+      } else {
+        _selectedAccountIds.removeAll(conversation.memberIds ?? []);
+      }
+    });
+  }
+
+  void clearSelectedIds() {
+    setStateIfMounted(() {
+      _selectedAccountIds.clear();
+    });
   }
 
   Future<void> _onRefresh() async => (_selectedTab == 0) ?
     _recentPageKey.currentState?.refresh() : _allUsersPageKey.currentState?.refresh();
 
-  bool get _hasSelectedAccounts => (_selectedTab == 0) ?
-    (_recentPageKey.currentState?.selectedConversationIds.isNotEmpty == true) :
-    (_allUsersPageKey.currentState?.selectedAccountIds.isNotEmpty == true);
+  bool get _hasSelectedAccounts => _selectedAccountIds.isNotEmpty;
 
 }